MODRS-4 - Refactoring (after build failure)
