+++ descriptors/ModuleDescriptor-template.json
@@ -1,12 +1,7 @@
 {
   "id": "@artifactId@-@version@",
   "name": "Remote storage API module",
-  "requires": [
-    {
-      "id": "inventory",
-      "version": "10.7"
-    }
-  ],
+  "requires": [],
   "provides": [
     {
       "id": "remote-storage-configurations",
@@ -87,14 +82,7 @@
       "handlers": [
         {
           "methods": ["POST"],
-          "pathPattern": "/_/tenant",
-          "modulePermissions": [
-            "users.collection.get",
-            "users.item.post",
-            "users.item.put",
-            "login.item.post",
-            "perms.users.item.post"
-          ]
+          "pathPattern": "/_/tenant"
         }, {
           "methods": ["DELETE"],
           "pathPattern": "/_/tenant"
@@ -186,6 +174,8 @@
       },
       { "name": "DB_HOST", "value": "postgres" },
       { "name": "DB_PORT", "value": "5432" },
+      { "name": "KAFKA_HOST", "value": "kafka"},
+      { "name": "KAFKA_PORT", "value": "9092"},
       { "name": "DB_USERNAME", "value": "folio_admin" },
       { "name": "DB_PASSWORD", "value": "folio_admin" },
       { "name": "DB_DATABASE", "value": "okapi_modules" },
