+++ docker/docker-compose.yml
@@ -1,0 +1,35 @@
+version: "3.8"
+
+services:
+
+  zookeeper:
+    image: zookeeper:3.6.2
+    container_name: zookeeper
+    ports:
+      - 2181:2181
+      - 8080:8080
+    environment:
+      ZOOKEEPER_SERVER_ID: 1
+      ZOO_PORT_NUMBER: 2181
+
+  kafka:
+    image: wurstmeister/kafka:2.13-2.6.0
+    container_name: kafka
+    hostname: kafka
+    depends_on:
+      - zookeeper
+    ports:
+      - 9092:9092
+    environment:
+      KAFKA_BROKER_ID: 1
+      KAFKA_LISTENERS: INSIDE://:9091,OUTSIDE://:9092
+      KAFKA_ADVERTISED_LISTENERS: INSIDE://:9091,OUTSIDE://kafka:9092
+      KAFKA_ADVERTISED_HOST_NAME: kafka
+      KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
+      KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: INSIDE:PLAINTEXT,OUTSIDE:PLAINTEXT
+      KAFKA_INTER_BROKER_LISTENER_NAME: INSIDE
+      KAFKA_MESSAGE_MAX_BYTES: 1000000
+      KAFKA_AUTO_CREATE_TOPICS_ENABLE: "false"
+      MIN_INSYNC_REPLICAS: 2
+      KAFKA_CREATE_TOPICS: >-
+        inventory.item:4:1
