--- src/main/resources/application.properties
@@ -1,23 +1,0 @@
-server.port=8081
-spring.application.name=mod-remote-storage
-#
-spring.datasource.url=jdbc:postgresql://localhost:5432/okapi_modules
-#
-spring.liquibase.enabled=true
-spring.liquibase.changeLog=classpath:db/changelog/changelog-master.xml
-#
-spring.jpa.hibernate.ddl-auto=none
-spring.jpa.properties.physical-strategy=org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
-spring.jpa.properties.implicit-strategy=org.springframework.boot.orm.jpa.hibernate.SpringImplicitNamingStrategy
-
-#
-spring.jpa.show-sql=false
-spring.jpa.properties.hibernate.format_sql=true
-#
-spring.jackson.default-property-inclusion=NON_NULL
-#
-folio.tenant.validation.enabled=true
-#
-application.kafka.listener.events.concurrency=1
-application.kafka.listener.events.topics=inventory.item
-application.kafka.listener.events.group-id=mod-remote-storage-items-group
