+++ src/test/java/org/folio/rs/TestSuite.java
@@ -2,7 +2,7 @@ package org.folio.rs;
 
 import org.folio.rs.controller.ConfigurationsTest;
 import org.folio.rs.controller.LocationMappingsTest;
-import org.folio.rs.integration.KafkaIntegrationTest;
+import org.folio.rs.service.AccessionQueueServiceTest;
 import org.folio.rs.service.SecurityManagerServiceTest;
 import org.junit.jupiter.api.Nested;
 
@@ -25,7 +25,7 @@ public class TestSuite {
   }
 
   @Nested
-  class KafkaIntegrationTestNested extends KafkaIntegrationTest {
+  class AccessionQueueServiceTestNested extends AccessionQueueServiceTest {
 
   }
 }
