--- src/test/resources/application-test.properties
@@ -1,26 +1,0 @@
-server.port=8081
-spring.application.name=mod-remote-storage
-#
-spring.liquibase.enabled=true
-spring.liquibase.changeLog=classpath:db/changelog/changelog-master.xml
-#
-spring.jpa.hibernate.ddl-auto=none
-#
-spring.jpa.show-sql=false
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.show-sql=false
-
-spring.jpa.properties.physical-strategy=org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
-spring.jpa.properties.implicit-strategy=org.springframework.boot.orm.jpa.hibernate.SpringImplicitNamingStrategy
-
-#
-spring.jackson.default-property-inclusion=NON_EMPTY
-#
-folio.tenant.validation.enabled=false
-
-spring.flyway.enabled=false
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
-
-application.kafka.listener.events.concurrency=1
-application.kafka.listener.events.topics=inventory.item
-application.kafka.listener.events.group-id=mod-remote-storage-items-group
